Byte order reversal shouldn't assume size_t is 64 bit.
This fixes compilation on 32 bit i386 with -WError.

 ../lib/crypto/aes_gcm_128.c:213:2: error: right shift count >= width of type [-Werror]
 ../lib/crypto/aes_gcm_128.c:213:2: error: left shift count >= width of type [-Werror]

+++ lib/util/byteorder.h
@@ -164,7 +164,7 @@ static __inline__ void st_le32(uint32_t *addr, const uint32_t val)
 /* now the reverse routines - these are used in nmb packets (mostly) */
 #define SREV(x) ((((x)&0xFF)<<8) | (((x)>>8)&0xFF))
 #define IREV(x) ((SREV(x)<<16) | (SREV((x)>>16)))
-#define BREV(x) ((IREV(x)<<32) | (IREV((x)>>32)))
+#define BREV(x) ((IREV((uint64_t)x)<<32) | (IREV(((uint64_t)x)>>32)))
 
 #define RSVAL(buf,pos) SREV(SVAL(buf,pos))
 #define RSVALS(buf,pos) SREV(SVALS(buf,pos))
